project/remotecontrol/pyproject.toml
[tool.poetry]
name = "remotecontrol"
version = "0.1.0"
description = ""
authors = ["Андрей Ростовцев <Electron4ik123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
drf-spectacular = "^0.24.2"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
pytest-coverage = "^0.0"
pytest = "^7.2.0"
drf-spectacular = "^0.24.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
